Cree los metodos para validar el usuario con el SharedServer
